Add basic color palette and altair configuration
