+++ great_expectations/rule_based_profiler/types/__init__.py
@@ -1,3 +1,5 @@
+from .altair_configuration import ALTAIR_CONFIGURATION
+
 from .attributes import Attributes  # isort:skip
 from .builder import Builder  # isort:skip
 from .domain import (  # isort:skip
