+++ great_expectations/types/color_palette.py
@@ -1,0 +1,13 @@
+from typing import Dict
+
+COLOR_PALETTE: Dict[str, Dict[str, str]] = {
+    "colors": {
+        "green_1": "#00C2A4",
+        "pink": "#FD5383",
+        "purple": "#8784FF",
+        "blue_1": "#1B2A4D",
+        "blue_2": "#384B74",
+        "blue_3": "#8699B7",
+    },
+    "category": ["#1B2A4D", "#00C2A4", "#8784FF", "#FD5383", "#8699B7"],
+}
